Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-typeMismatch-expected.txt


PASS [INPUT in EMAIL status] The value is empty
PASS [INPUT in EMAIL status] The value is a valid email address
PASS [INPUT in EMAIL status] The value is a valid email address with some white spaces.
FAIL [INPUT in EMAIL status] The value is not an email address assert_true: The validity.typeMismatch should be true, when control is disabled. expected true got false
FAIL [INPUT in EMAIL status] The value contains multiple email addresses assert_true: The validity.typeMismatch should be true, when control is disabled. expected true got false
PASS [INPUT in EMAIL status] The value is valid email addresses
FAIL [INPUT in EMAIL status] The value contains invalid separator assert_true: The validity.typeMismatch should be true, when control is disabled. expected true got false
PASS [INPUT in URL status] The value is empty
PASS [INPUT in URL status] The value is a valid url
PASS [INPUT in URL status] The value is a valid url with some white spaces.
FAIL [INPUT in URL status] The value is not an url assert_true: The validity.typeMismatch should be true, when control is disabled. expected true got false
